Sagar_Dataharvesting/db_indices_management/requirements.txt
fastapi==0.115.4
mysql_connector_repackaged==0.3.1
pydantic==2.9.2
